feat: audit logs, license soft-gate, and validation hardening for proxied services
Backend:
- emit audit logs on proxied service create/update/delete
- soft-gate the dashboard proxied-service methods on the secretsBrokering license
  (return empty rather than throwing, so the shared overview isn't broken)
- move host-pattern + credential validation schemas into proxied-service-schemas.ts
  and switch the backend regexes to RE2 (codebase convention)
- de-duplicate the sanitized response schema (shared base + credential picks)

Frontend:
- mirror the host-pattern grammar + 255 cap in the form schema (inline errors)
- surface the backend error message on save failure instead of a generic toast
- gate the dashboard includeProxiedServices flag on the subscription
